requirements.txt
loguru
python-dotenv
requests
aiogram
numpy
pandas
vkbottle